Add converter package for tar build
Introduce a new nydus image build mode named tar build.

In order to correspond to the OCI image layer, we need a new build mode, that is
to convert the OCI-formatted tar stream directly into a nydus-formatted tar stream.

The nydus tar stream contains blob and bootstrap files, and the workflow is divided
into two steps, Convert and Merge.

Convert: converts each layer of the OCI tar stream to a nydus tar stream.
Merge: reads the bootstrap files from each layer of the nydus tar stream and merges
them into a final bootstrap.

This package will be imported by acceleration-service and buildkit to support the
conversion or build of nydus image.

Signed-off-by: Yan Song <imeoer@linux.alibaba.com>